restructure heuristic type definitions to be local to app config
